refactor tests for artifact store implementations to be parametric over artifact type and store type
